move lexicon-specific list of files
